Added RsvpPolicy

+++ app/Http/Controllers/RsvpController.php
@@ -14,6 +14,8 @@ class RsvpController extends Controller
 {
     public function store(Request $request, Event $event): RedirectResponse
     {
+        $this->authorize('create', Rsvp::class);
+
         $request->validate([
             'rsvp_response' => 'required',
         ]);
@@ -28,6 +30,8 @@ class RsvpController extends Controller
 
     public function update(Request $request, Event $event, Rsvp $rsvp): RedirectResponse
     {
+        $this->authorize('update', $rsvp);
+
         $request->validate([
             'rsvp_response' => 'required',
         ]);
@@ -41,6 +45,8 @@ class RsvpController extends Controller
 
     public function destroy(Event $event, Rsvp $rsvp): RedirectResponse
     {
+        $this->authorize('delete', $rsvp);
+
         $rsvp->delete();
 
         return redirect()->route('events.show', [$event])->with(['status' => 'RSVP deleted.']);
